+++ app/src/main/java/ru/livetyping/zarina/domain/common/PriceRange.kt
@@ -1,8 +1,8 @@
 package ru.livetyping.zarina.domain.common
 
 data class PriceRange(
-    val min: Long,
-    val max: Long,
+    val min: Int,
+    val max: Int,
 ) {
     init {
         require(max >= min) { "\"max\" $max must be equal to or larger than \"min\" $min" }
@@ -10,6 +10,6 @@ data class PriceRange(
 
     companion object {
         val EMPTY: PriceRange
-            get() = PriceRange(min = 0L, max = Long.MAX_VALUE)
+            get() = PriceRange(min = 0, max = Int.MAX_VALUE)
     }
 }
